pyproject.toml
[tool.poetry]
name = "babichjacob-oneshot-channel"
version = "0.1.0"
description = "TODO"
license = "MIT"
authors = ["J or Jacob Babich <jacobbabichpublic+git@gmail.com>"]
readme = "README.md"
repository = "https://github.com/babichjacob/python-oneshot-channel"
packages = [{include = "oneshot_channel"}]

[tool.poetry.dependencies]
python = ">=3.10"
babichjacob-option-and-result = "^0.1.2"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
mypy = "^0.991"
pylint = "^2.15.10"

[tool.poetry.group.test.dependencies]
pytest = "^6"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
